requirements.txt
flask
gradio
torch
torchvision
pytorch_lightning
numpy
Pillow
matplotlib
wandb
huggingface_hub